DCELs

Documentation for DCELs.

DCELs.AbstractEdge

DCELs.AbstractFace

DCELs.AbstractGeometry

DCELs.AbstractTessellation

DCELs.AbstractVertex

DCELs.Edge

DCELs.Edge

DCELs.Face

DCELs.Face

DCELs.GeometryError

DCELs.Tessellation

DCELs.Tessellation

DCELs.Vertex

DCELs.Vertex

DCELs.Vertex

DCELs.connect

DCELs.edge_loop

DCELs.edge_path

DCELs.edges

DCELs.edges

DCELs.faces

DCELs.faces

DCELs.faces

DCELs.find_outer_face

DCELs.fix_faces!

DCELs.squeeze_edge

DCELs.vertices

DCELs.vertices

DCELs.AbstractEdge — Type

AbstractEdge

Abstract supertype for all edge objects.

DCELs.AbstractFace — Type

AbstractFace

Abstract supertype for all face objects.

DCELs.AbstractGeometry — Type

AbstractGeometry

Abstract supertype for all geometry objects.

DCELs.AbstractTessellation — Type

AbstractTessellation

Abstract supertype for all tessellations.

DCELs.AbstractVertex — Type

AbstractVertex

Abstract supertype for all vertex objects.

DCELs.Edge — Type

Edge <: AbstractEdge

Edge between two vertices.

Fields

id::String: Unique ID.

next::Union{Edge, Nothing}`: Next edge alongside the corresponding face

(counterclockwise).

prev::Union{Edge, Nothing}`: Previous edge alongside the corresponding face

(counterclockwise).

twin::Union{Edge, Nothing}`: Twin edge.

prev: Vertex that the edge points to.

face: Face bounded by the edge.

DCELs.Edge — Method

Edge()

Create an Edge with a random UUID4 ID prefixed by "E-", and all other fields set to nothing.

DCELs.Face — Type

Face <: AbstractFace

Area in the plane bounded by edges.

Fields

id::String: Unique ID.

edge::Union{Edge, Nothing}: Arbitrary edge that bounds this face.

DCELs.Face — Method

Face()

Create a Face with a random UUID4 ID prefixed by "F-", and edge=nothing.

DCELs.GeometryError — Type

GeometryError(msg::String)

Geometry is inconsistent.

DCELs.Tessellation — Type

Tessellation <: AbstractTessellation

Collection of vertices, edges and faces.

Fields

edges::Set{Edge}: Set of edges belonging to the tessellation.

vertices::Set{Vertex}: Set of vertices.

faces::Set{Face}: Set of faces.

DCELs.Tessellation — Method

Tessellation()

Create a Tessellation with no edges, vertices or faces.

DCELs.Vertex — Type

Vertex <: AbstractVertex

2D point in the plane.

Fields

id::String: Unique ID.

x::Real: x coordinate.

y::Real: y coordinate.

edge::Union{Edge, Nothing}: Arbitrary edge that points towards this vertex.

DCELs.Vertex — Method

Vertex(x::Real, y::Real)

Create a Vertex with a random UUID4 ID prefixed by "V-", with coordinates (x, y), and edge=nothing.

DCELs.Vertex — Method

Vertex()

Create a Vertex with a random UUID4 ID prefixed by "V-", with coordinates (NaN, NaN), and edge=nothing.

DCELs.connect — Method

connect(args...; kwargs...)

Connect two vertices.

Arguments

tess::AbstractTessellation: Tessellation to update.

u::AbstractVertex: First vertex to connect.

v::AbstractVertex: Second vertex to connect.

Keywords

split::Union{Bool, Nothing}=nothing: Whether connecting u and v will create a new

face. If nothing, it will be determined automatically, at the cost of more operations.

Returns

h: The newly created edge from u to v.`

DCELs.edge_loop — Method

edge_loop(start::AbstractEdge; reverse::Bool=false)

Yield all edges in the list [start, start.next, start.next.next, ..., start.prev]. Ifreverse=true, this will instead be[start, start.prev, ..., start.next]`.

DCELs.edge_path — Method

edge_path(start::AbstractEdge, finish::AbstractEdge; reverse::Bool=false)

Yield all edges in the list [start, start.next, start.next.next, ..., finish.prev]. If start and finish do not bound the same face, this will never terminate.

If reverse=true, the order will be [start, start.prev, ..., finish.next].

DCELs.edges — Method

edges(face::AbstractFace; reverse::Bool=false)

Yield all edges that bound face, in counterclockwise order (or clockwise if reverse=true).

DCELs.edges — Method

edges(vertex::AbstractVertex; reverse::Bool=false)

Yield all edges that point to vertex, in counterclockwise order (or clockwise if reverse=true).

DCELs.faces — Method

faces(edge::AbstractEdge; reverse::Bool=false)

Yield both faces that edge touches, starting by the left one (or right if reverse=true).

DCELs.faces — Method

faces(face::AbstractFace; reverse::Bool=false)

Yield all faces that share an edge with face, in counterclockwise order (or clockwise if reverse=true).

DCELs.faces — Method

faces(vertex::AbstractVertex; reverse::Bool=false)

Yield all of the faces that touch vertex, in counterclockwise order (or clockwise if reverse=true).

DCELs.find_outer_face — Method

find_outer_face(tess::AbstractTessellation)

Return the unbounded face of tess.

DCELs.fix_faces! — Method

fix_faces!(tess::AbstractTessellation)

Create the missing faces between the edges of tess. It is assumed that the prev and next fields are consistent.

DCELs.squeeze_edge — Method

squeeze_edge(vertex::AbstractVertex, edge::AbstractEdge)

Returns an edge h such that h.dest = vertex, and if you rotate edge counterclockwise around vertex, the first edge you'll meet is h.

DCELs.vertices — Method

vertices(edge::AbstractEdge; reverse::Bool=false)

Yield both endpoints of edge, starting at the destination (or origin if reverse=true).

DCELs.vertices — Method

vertices(face::AbstractFace; reverse::Bool=false)

Yield all vertices that touch face, in counterclockwise order (or clockwise if reverse=true).